security/libressl: downgrade because of breakage
The new LibreSSL version breaks a number of things, including some
py-* ports because some functionality was removed.

The proposed patches to py-cryptography do not work, it doesn't even
build with LibreSSL 2.7.x (PR 226906), and this breakage pulls things
like py-certbot with it as collateral damage.

Downgrading to 2.6.x for now, until everything catches up.